Refactor: organized the additional Login code
