+++ foodbook_app/lib/bloc/login_bloc/auth_state.dart
